combined PageRequest.cs & WebService.cs into Startup.cs to improve performance

+++ Server.cs
@@ -20,7 +20,7 @@ public class Server
     public Utility.Util Util = new Utility.Util();
     public enumEnvironment environment = enumEnvironment.development;
     public DateTime serverStart = DateTime.Now;
-    public int requestCount = 0;
+    public double requestCount = 0;
     public float requestTime = 0;
     public string nameSpace = "";
     public string sqlActive = "";
@@ -68,7 +68,7 @@ public class Server
     /// <summary>
     /// Loads a file from cache. If the file hasn't been cached yet, then load file from a drive.
     /// </summary>
-    /// <param name="filename"></param>
+    /// <param name="filename">The relevant path to the file</param>
     /// <param name="noDevEnvCache">If true, it will not load a file from cache if the app is running in a development environment. Instead, it will always load the file from a drive.</param>
     /// <param name="noCache">If true, will not save to cache, but will instead load file from disk every time</param>
         /// <returns></returns>
